串口通讯测试修改；

+++ USER/motor.h
@@ -19,7 +19,7 @@
 //#define Motor_InitSpeed                0x57e3
 
 #define Motor_Move_MAX_LEN      80//97//mm
-#define Motor_Move_MAX_STEP     Motor_Move_MAX_LEN*Motor_StepsPerum  //    
+#define Motor_Move_MAX_STEP     (u32)(Motor_Move_MAX_LEN*Motor_StepsPerum)  //    
 //#define Motor_Timer_PSC         1
 //#define Motor_Timer_CLK         (rcc_clocks.PCLK2_Frequency/Motor_Timer_PSC)
 //#define DEF_FREQ_MIN            2000
